% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_imports.R
\name{clear_imports}
\alias{clear_imports}
\title{Clear the \pkg{pysty} Environments}
\usage{
clear_imports(..., pos = 1, envir = as.environment(pos))
}
\arguments{
\item{pos}{Where to do the assignment.  If no arguments are passed all
enviroments will be cleared.}

\item{envir}{The \code{\link[base]{environment}} to use.}

\item{\ldots}{Quoted package names to remove from the \pkg{pysty} environments
as well as any functions with those package name(s) from the global enviroment.}
}
\description{
Clear the imports from the \pkg{pysty} environments as well as any functions
with those package name(s) from the global enviroment.
}
\examples{
\dontrun{
add_imports('import dplyr as dp', 'from lattice import xyplot, qq')
get_imports(); get_imports_functions();ls(pattern='^(qq|xyplot)$')
clear_imports()
get_imports(); get_imports_functions();ls(pattern='^(qq|xyplot)$')
}
}
\seealso{
\code{\link[base]{assign}},
\code{\link[base]{get}}
}

