% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colons.R
\name{::}
\alias{::}
\alias{:::}
\title{Double & Triple Colon Operators}
\usage{
"::"(x, y)

":::"(x, y)
}
\arguments{
\item{x}{A package or \pkg{pysty} alias.}

\item{y}{A function (exported or non-exported) in that package.}
}
\description{
Replaces the \pkg{base} colon operators with \pkg{pysty} varients that allow for
aliased package names.
}
\examples{
\dontrun{
add_imports('import lattice as lat')
lat::xyplot
lat:::wireframe.matrix
}
}
\seealso{
\code{\link[base]{::}}
}

