% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imports.R
\name{get_imports}
\alias{get_imports}
\title{View Imported Packages}
\usage{
get_imports(...)
}
\arguments{
\item{\ldots}{ignored.}
}
\value{
Returns a data.frame of packages and aliases.
}
\description{
View imported packages (using \pkg{pysty}) and aliases.
}
\examples{
\dontrun{
add_imports('import lattice as lat')
get_imports()
}
}
