% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imports_functions.R
\name{get_imports_functions}
\alias{get_imports_functions}
\title{Get Functions Imported From Packages}
\usage{
get_imports_functions(...)
}
\arguments{
\item{\ldots}{ignored.}
}
\value{
Returns a data.frame of packages and functions.
}
\description{
Look at the functions that are imported into the global environment from
packages.
}
\examples{
\dontrun{
add_imports('from lattice import xyplot, qq')
get_imports_functions()
}
}

