% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pysty-package.R
\docType{package}
\name{pysty}
\alias{pysty}
\alias{package-pysty}
\alias{pysty-package}
\title{\pkg{Python} Style Package Importing}
\description{
\pkg{Python} style packages importing using the common forms of:
\code{import PACKAGE}, \code{import PACKAGE as ALIAS}, or
\code{from PACKAGE import FUN1, FUN2, FUN_N}.
}
